refactor(handler): restructure handler files
