+++ internal/adapter/handler/http/user.go
@@ -1,9 +1,6 @@
 package handler
 
 import (
-	"net/http"
-	"time"
-
 	"github.com/bagashiz/go-pos/internal/core/domain"
 	"github.com/bagashiz/go-pos/internal/core/port"
 	"github.com/gin-gonic/gin"
@@ -18,26 +15,6 @@ type UserHandler struct {
 func NewUserHandler(svc port.UserService) *UserHandler {
 	return &UserHandler{
 		svc,
-	}
-}
-
-// userResponse represents a user response body
-type userResponse struct {
-	ID        uint64    `json:"id" example:"1"`
-	Name      string    `json:"name" example:"John Doe"`
-	Email     string    `json:"email" example:"test@example.com"`
-	CreatedAt time.Time `json:"c0eated_at" example:"1970-01-01T00:00:00Z"`
-	UpdatedAt time.Time `json:"updated_at" example:"1970-01-01T00:00:00Z"`
-}
-
-// newUserResponse is a helper function to create a response body for handling user data
-func newUserResponse(user *domain.User) userResponse {
-	return userResponse{
-		ID:        user.ID,
-		Name:      user.Name,
-		Email:     user.Email,
-		CreatedAt: user.CreatedAt,
-		UpdatedAt: user.UpdatedAt,
 	}
 }
 
@@ -56,17 +33,17 @@ type registerRequest struct {
 //	@Accept			json
 //	@Produce		json
 //	@Param			registerRequest	body		registerRequest	true	"Register request"
-//	@Success		201				{object}	userResponse	"User created"
-//	@Failure		400				{object}	response		"Validation error"
-//	@Failure		401				{object}	response		"Unauthorized error"
-//	@Failure		404				{object}	response		"Data not found error"
-//	@Failure		409				{object}	response		"Data conflict error"
-//	@Failure		500				{object}	response		"Internal server error"
+//	@Success		200				{object}	userResponse	"User created"
+//	@Failure		400				{object}	errorResponse	"Validation error"
+//	@Failure		401				{object}	errorResponse	"Unauthorized error"
+//	@Failure		404				{object}	errorResponse	"Data not found error"
+//	@Failure		409				{object}	errorResponse	"Data conflict error"
+//	@Failure		500				{object}	errorResponse	"Internal server error"
 //	@Router			/users [post]
 func (uh *UserHandler) Register(ctx *gin.Context) {
 	var req registerRequest
 	if err := ctx.ShouldBindJSON(&req); err != nil {
-		errorResponse(ctx, http.StatusBadRequest, err)
+		validationError(ctx, err)
 		return
 	}
 
@@ -84,7 +61,7 @@ func (uh *UserHandler) Register(ctx *gin.Context) {
 
 	rsp := newUserResponse(&user)
 
-	successResponse(ctx, http.StatusCreated, rsp)
+	handleSuccess(ctx, rsp)
 }
 
 // listUsersRequest represents the request body for listing users
@@ -100,18 +77,18 @@ type listUsersRequest struct {
 //	@Tags			Users
 //	@Accept			json
 //	@Produce		json
-//	@Param			skip	query		uint64		true	"Skip"
-//	@Param			limit	query		uint64		true	"Limit"
-//	@Success		200		{object}	response	"Users displayed"
-//	@Failure		400		{object}	response	"Validation error"
-//	@Failure		500		{object}	response	"Internal server error"
+//	@Param			skip	query		uint64			true	"Skip"
+//	@Param			limit	query		uint64			true	"Limit"
+//	@Success		200		{object}	meta			"Users displayed"
+//	@Failure		400		{object}	errorResponse	"Validation error"
+//	@Failure		500		{object}	errorResponse	"Internal server error"
 //	@Router			/users [get]
 func (uh *UserHandler) ListUsers(ctx *gin.Context) {
 	var req listUsersRequest
 	var usersList []userResponse
 
 	if err := ctx.ShouldBindQuery(&req); err != nil {
-		errorResponse(ctx, http.StatusBadRequest, err)
+		validationError(ctx, err)
 		return
 	}
 
@@ -129,7 +106,7 @@ func (uh *UserHandler) ListUsers(ctx *gin.Context) {
 	meta := newMeta(total, req.Limit, req.Skip)
 	rsp := toMap(meta, usersList, "users")
 
-	successResponse(ctx, http.StatusOK, rsp)
+	handleSuccess(ctx, rsp)
 }
 
 // getUserRequest represents the request body for getting a user
@@ -146,14 +123,14 @@ type getUserRequest struct {
 //	@Produce		json
 //	@Param			id	path		uint64			true	"User ID"
 //	@Success		200	{object}	userResponse	"User displayed"
-//	@Failure		400	{object}	response		"Validation error"
-//	@Failure		404	{object}	response		"Data not found error"
-//	@Failure		500	{object}	response		"Internal server error"
+//	@Failure		400	{object}	errorResponse	"Validation error"
+//	@Failure		404	{object}	errorResponse	"Data not found error"
+//	@Failure		500	{object}	errorResponse	"Internal server error"
 //	@Router			/users/{id} [get]
 func (uh *UserHandler) GetUser(ctx *gin.Context) {
 	var req getUserRequest
 	if err := ctx.ShouldBindUri(&req); err != nil {
-		errorResponse(ctx, http.StatusBadRequest, err)
+		validationError(ctx, err)
 		return
 	}
 
@@ -165,7 +142,7 @@ func (uh *UserHandler) GetUser(ctx *gin.Context) {
 
 	rsp := newUserResponse(user)
 
-	successResponse(ctx, http.StatusOK, rsp)
+	handleSuccess(ctx, rsp)
 }
 
 // updateUserRequest represents the request body for updating a user
@@ -186,23 +163,23 @@ type updateUserRequest struct {
 //	@Param			id					path		uint64				true	"User ID"
 //	@Param			updateUserRequest	body		updateUserRequest	true	"Update user request"
 //	@Success		200					{object}	userResponse		"User updated"
-//	@Failure		400					{object}	response			"Validation error"
-//	@Failure		401					{object}	response			"Unauthorized error"
-//	@Failure		404					{object}	response			"Data not found error"
-//	@Failure		500					{object}	response			"Internal server error"
+//	@Failure		400					{object}	errorResponse		"Validation error"
+//	@Failure		401					{object}	errorResponse		"Unauthorized error"
+//	@Failure		404					{object}	errorResponse		"Data not found error"
+//	@Failure		500					{object}	errorResponse		"Internal server error"
 //	@Router			/users/{id} [put]
 //	@Security		BearerAuth
 func (uh *UserHandler) UpdateUser(ctx *gin.Context) {
 	var req updateUserRequest
 	if err := ctx.ShouldBindJSON(&req); err != nil {
-		errorResponse(ctx, http.StatusBadRequest, err)
+		validationError(ctx, err)
 		return
 	}
 
 	idStr := ctx.Param("id")
 	id, err := stringToUint64(idStr)
 	if err != nil {
-		errorResponse(ctx, http.StatusBadRequest, err)
+		validationError(ctx, err)
 		return
 	}
 
@@ -222,7 +199,7 @@ func (uh *UserHandler) UpdateUser(ctx *gin.Context) {
 
 	rsp := newUserResponse(&user)
 
-	successResponse(ctx, http.StatusOK, rsp)
+	handleSuccess(ctx, rsp)
 }
 
 // deleteUserRequest represents the request body for deleting a user
@@ -237,18 +214,18 @@ type deleteUserRequest struct {
 //	@Tags			Users
 //	@Accept			json
 //	@Produce		json
-//	@Param			id	path		uint64		true	"User ID"
-//	@Success		200	{object}	response	"User deleted"
-//	@Failure		400	{object}	response	"Validation error"
-//	@Failure		401	{object}	response	"Unauthorized error"
-//	@Failure		404	{object}	response	"Data not found error"
-//	@Failure		500	{object}	response	"Internal server error"
+//	@Param			id	path		uint64			true	"User ID"
+//	@Success		200	{object}	response		"User deleted"
+//	@Failure		400	{object}	errorResponse	"Validation error"
+//	@Failure		401	{object}	errorResponse	"Unauthorized error"
+//	@Failure		404	{object}	errorResponse	"Data not found error"
+//	@Failure		500	{object}	errorResponse	"Internal server error"
 //	@Router			/users/{id} [delete]
 //	@Security		BearerAuth
 func (uh *UserHandler) DeleteUser(ctx *gin.Context) {
 	var req deleteUserRequest
 	if err := ctx.ShouldBindUri(&req); err != nil {
-		errorResponse(ctx, http.StatusBadRequest, err)
+		validationError(ctx, err)
 		return
 	}
 
@@ -258,5 +235,5 @@ func (uh *UserHandler) DeleteUser(ctx *gin.Context) {
 		return
 	}
 
-	successResponse(ctx, http.StatusOK, nil)
+	handleSuccess(ctx, nil)
 }
